[wrangler] Move queues/ commands to src/commands/queues/ (Phase 2d)
