+++ packages/wrangler/src/commands/queues/cli/commands/consumer/worker/add.ts
@@ -1,6 +1,6 @@
 import { CommandLineArgsError } from "@cloudflare/workers-utils";
-import { createCommand } from "../../../../../core/create-command";
-import { logger } from "../../../../../logger";
+import { createCommand } from "../../../../../../core/create-command";
+import { logger } from "../../../../../../logger";
 import { postConsumer } from "../../../../client";
 
 export const queuesConsumerAddCommand = createCommand({
