+++ packages/wrangler/src/commands/queues/cli/commands/consumer/worker/index.ts
@@ -1,0 +1,9 @@
+import { createNamespace } from "../../../../../../core/create-command";
+
+export const queuesConsumerWorkerNamespace = createNamespace({
+	metadata: {
+		description: "Configure Queue Worker Consumers",
+		owner: "Product: Queues",
+		status: "stable",
+	},
+});
