+++ packages/wrangler/src/commands/queues/cli/commands/consumer/worker/remove.ts
@@ -1,5 +1,5 @@
-import { createCommand } from "../../../../../core/create-command";
-import { logger } from "../../../../../logger";
+import { createCommand } from "../../../../../../core/create-command";
+import { logger } from "../../../../../../logger";
 import { deleteWorkerConsumer } from "../../../../client";
 
 export const queuesConsumerRemoveCommand = createCommand({
