+++ packages/wrangler/src/commands/queues/cli/commands/delete.ts
@@ -1,5 +1,5 @@
-import { createCommand } from "../../../core/create-command";
-import { logger } from "../../../logger";
+import { createCommand } from "../../../../core/create-command";
+import { logger } from "../../../../logger";
 import { deleteQueue } from "../../client";
 
 export const queuesDeleteCommand = createCommand({
