+++ packages/wrangler/src/commands/queues/cli/commands/subscription/delete.ts
@@ -1,6 +1,6 @@
-import { createCommand } from "../../../../core/create-command";
-import { confirm } from "../../../../dialogs";
-import { logger } from "../../../../logger";
+import { createCommand } from "../../../../../core/create-command";
+import { confirm } from "../../../../../dialogs";
+import { logger } from "../../../../../logger";
 import {
 	deleteEventSubscription,
 	getEventSubscriptionForQueue,
