+++ packages/wrangler/src/commands/queues/cli/commands/subscription/list.ts
@@ -1,5 +1,5 @@
-import { createCommand } from "../../../../core/create-command";
-import { logger } from "../../../../logger";
+import { createCommand } from "../../../../../core/create-command";
+import { logger } from "../../../../../logger";
 import { listEventSubscriptions } from "../../../client";
 import { getSourceResource, getSourceType } from "./utils";
