+++ packages/wrangler/src/commands/queues/cli/commands/subscription/update.ts
@@ -1,6 +1,6 @@
 import { UserError } from "@cloudflare/workers-utils";
-import { createCommand } from "../../../../core/create-command";
-import { logger } from "../../../../logger";
+import { createCommand } from "../../../../../core/create-command";
+import { logger } from "../../../../../logger";
 import {
 	getEventSubscriptionForQueue,
 	updateEventSubscription,
